Move consingment submission fails test to the contatct information step
